+++ src/Services/ServiceCaller.php
@@ -3,13 +3,14 @@
 namespace PerfectOblivion\ActionServiceResponder\Services;
 
 use Illuminate\Contracts\Bus\Dispatcher;
-use Illuminate\Support\Collection;
 use PerfectOblivion\ActionServiceResponder\Services\Contracts\ShouldQueueService;
 use PerfectOblivion\ActionServiceResponder\Services\Exceptions\ServiceHandlerMethodException;
-use PerfectOblivion\ActionServiceResponder\Services\Service;
+use PerfectOblivion\ActionServiceResponder\Services\Traits\DisablesAutorun;
 
 class ServiceCaller extends AbstractServiceCaller
 {
+    use DisablesAutorun;
+
     /** @var \PerfectOblivion\ActionServiceResponder\Services\Service */
     protected $resolvedService;
 
@@ -17,105 +18,77 @@ class ServiceCaller extends AbstractServiceCaller
      * Call a service through its appropriate handler.
      *
      * @param  string  $service
-     * @param  array  $params
-     * @param  array  $supplementals
+     * @param  array  $parameters
+     * @param  array  $supplementalParameters
      *
      * @return mixed
      */
-    public function call(string $service, array $params, array $supplementals = [])
+    public function call(string $service, array $parameters, array $supplementalParameters = [])
     {
-        $this->prepareService($service, $supplementals);
-        $this->validateData($params);
+        $this->prepareService($service, $supplementalParameters);
+        $this->validateData($parameters);
 
-        return $this->shouldQueueService($this->resolvedService)
-            ? $this->dispatchService(
-                $params,
-                $this->resolvedService->getData(),
-                $this->resolvedService->isValidated(),
-                $this->resolvedService->getSupplementals(),
-                get_object_vars($this->resolvedService),
-            )
-            : $this->resolvedService->run($params);
+        return $this->shouldQueueService() ? $this->dispatchService($parameters) : $this->resolvedService->run($parameters);
     }
 
     /**
-     * Push the service call to the queue..
+     * Push the service call to the queue.
      *
      * @param  string  $service
-     * @param  array  $params
-     * @param  array  $supplementals
+     * @param  array  $parameters
+     * @param  array  $supplementalParameters
      *
      * @throws \PerfectOblivion\ActionServiceResponder\Exceptions\ServiceHandlerMethodException
      */
-    public function queue(string $service, array $params, array $supplementals = []): void
+    public function queue(string $service, array $parameters, array $supplementalParameters = []): void
     {
-        $this->prepareService($service, $supplementals);
-        $this->validateData($params);
-        $this->dispatchService(
-            $params,
-            $this->resolvedService->getData(),
-            $this->resolvedService->isValidated(),
-            $this->resolvedService->getSupplementals(),
-            get_object_vars($this->resolvedService),
-        );
+        $this->prepareService($service, $supplementalParameters);
+        $this->validateData($parameters);
+        $this->dispatchService($parameters);
     }
 
     /**
      * Prepare the Service to be run.
      *
      * @param  string  $service
-     * @param  array  $supplementals
+     * @param  array  $supplementalParameters
      */
-    private function prepareService(string $service, array $supplementals): void
+    private function prepareService(string $service, array $supplementalParameters): void
     {
         $this->confirmServiceHasHandler($service);
-        $this->temporarilyDisableAutorun();
+        $this->disableAutorun();
         $this->resolvedService = $this->container->make($service);
-        $this->resolvedService->buildSupplementals($supplementals);
+        $this->resolvedService->buildSupplementals($supplementalParameters);
     }
 
     /**
      * Call the service's validator if it exists and hasn't been called yet.
      *
-     * @param  array  $params
+     * @param  array  $parameters
      */
-    private function validateData(array $params = []): void
+    private function validateData(array $parameters = []): void
     {
         $validator = $this->resolvedService->getValidator();
 
         if ($validator) {
             $validator->service = $this->resolvedService;
             if (! $this->resolvedService->isValidated()) {
-                $this->resolvedService->setData($validator->validate($params));
+                $this->resolvedService->setData($validator->validate($parameters));
                 $this->resolvedService->setIsValidated(true);
             }
         } else {
-            $this->resolvedService->setData($params);
+            $this->resolvedService->setData($parameters);
         }
     }
 
     /**
      * Dispatch a fully initialized Service to the queue.
      *
-     * @param  array  $params
-     * @param  array  $data
-     * @param  bool  $validated
-     * @param  \Illuminate\Support\Collection  $supplementals
-     * @param  array  $props
+     * @param  array  $parameters
      */
-    private function dispatchService(array $params, array $data, bool $validated, Collection $supplementals, array $props): void
+    private function dispatchService(array $parameters): void
     {
-        resolve(Dispatcher::class)->dispatch(new QueuedService($this->resolvedService, $params, $data, $validated, $supplementals, $props));
-    }
-
-    /**
-     * Temporarily disable the service autorun.
-     */
-    private function temporarilyDisableAutorun(): void
-    {
-        $this->container->resolving(Service::class, function ($service, $app) {
-            $service->autorunIfEnabled = false;
-        });
+        resolve(Dispatcher::class)->dispatch(new QueuedService($this->resolvedService, $parameters));
     }
 
     /**
@@ -127,18 +100,14 @@ class ServiceCaller extends AbstractServiceCaller
      */
     private function confirmServiceHasHandler(string $service)
     {
-        if (! $this->hasHandler($service)) {
-            throw ServiceHandlerMethodException::notFound($service);
-        }
+        throw_unless($this->hasHandler($service), ServiceHandlerMethodException::notFound($service));
     }
 
     /**
      * Should the Service be queued?
-     *
-     * @param  \PerfectOblivion\ActionServiceResponder\Services\Service  $service
      */
-    private function shouldQueueService(Service $service): bool
+    private function shouldQueueService(): bool
     {
-        return $service instanceof ShouldQueueService;
+        return $this->resolvedService instanceof ShouldQueueService;
     }
 }