+++ src/Services/Traits/CallsServices.php
@@ -11,29 +11,29 @@ trait CallsServices
      * Call a service.
      *
      * @param  string  $service
-     * @param  array  $params
+     * @param  array  $parameters
      * @param  array  $supplementals
      *
      * @return mixed
      */
-    public function call(string $service, array $params, array $supplementals = [])
+    public function call(string $service, array $parameters, array $supplementals = [])
     {
-        return Container::getInstance()->make(ServiceCaller::class)->call($service, $params, $supplementals);
+        return Container::getInstance()->make(ServiceCaller::class)->call($service, $parameters, $supplementals);
     }
 
     /**
      * Push the service call to the queue..
      *
      * @param  string  $service
-     * @param  array  $params
+     * @param  array  $parameters
      * @param  array  $supplementals
      *
      * @throws \PerfectOblivion\ActionServiceResponder\Exceptions\ServiceHandlerMethodException
      *
      * @return mixed
      */
-    public function queue(string $service, array $params, array $supplementals = [])
+    public function queue(string $service, array $parameters, array $supplementals = [])
     {
-        return Container::getInstance()->make(ServiceCaller::class)->queue($service, $params, $supplementals);
+        return Container::getInstance()->make(ServiceCaller::class)->queue($service, $parameters, $supplementals);
     }
 }